+++ data-export/data-export-application/src/main/kotlin/eu/tib/orkg/prototype/export/comparisons/api/ExportUnpublishedComparisonUseCase.kt
@@ -1,0 +1,8 @@
+package eu.tib.orkg.prototype.export.comparisons.api
+
+import java.io.Writer
+
+interface ExportUnpublishedComparisonUseCase {
+    fun export(writer: Writer)
+    fun export(path: String?)
+}