+++ graph/graph-application/src/main/kotlin/eu/tib/orkg/prototype/contenttypes/adapter/output/datacite/DataCiteDoiServiceAdapter.kt
@@ -1,19 +1,18 @@
 package eu.tib.orkg.prototype.contenttypes.adapter.output.datacite
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo
-import com.fasterxml.jackson.annotation.JsonTypeName
 import com.fasterxml.jackson.databind.ObjectMapper
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.Attributes
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.Creator
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.Description
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.NameIdentifier
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.RelatedIdentifier
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.Rights
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.Subject
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.Title
-import eu.tib.orkg.prototype.contenttypes.adapter.output.datacite.DataCiteDoiServiceAdapter.RegisterDoiRequest.Type
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.Attributes
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.Creator
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.Description
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.NameIdentifier
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.RelatedIdentifier
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.Rights
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.Subject
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.Title
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson.Type
 import eu.tib.orkg.prototype.contenttypes.domain.model.DOI
 import eu.tib.orkg.prototype.contenttypes.spi.DoiService
+import eu.tib.orkg.prototype.datacite.json.DataCiteJson
 import eu.tib.orkg.prototype.statements.application.DOIServiceUnavailable
 import java.io.IOException
 import java.net.URI
@@ -64,98 +63,8 @@ class DataCiteDoiServiceAdapter(
         }
     }
 
-    @JsonTypeName(value = "data")
-    @JsonTypeInfo(include = JsonTypeInfo.As.WRAPPER_OBJECT, use = JsonTypeInfo.Id.NAME)
-    data class RegisterDoiRequest(
-        val attributes: Attributes,
-        val type: String = "dois"
-    ) {
-        data class Attributes(
-            val doi: String,
-            val event: String,
-            val creators: List<Creator>,
-            val titles: List<Title>,
-            val publicationYear: Int,
-            val subjects: List<Subject>,
-            val types: Type,
-            val relatedIdentifiers: List<RelatedIdentifier>,
-            val rightsList: List<Rights>,
-            val descriptions: List<Description>,
-            val url: URI,
-            val language: String = "en",
-            val publisher: String = "Open Research Knowledge Graph"
-        )
-
-        data class Creator(
-            val name: String,
-            val nameIdentifiers: List<NameIdentifier>,
-            val nameType: String = "Personal"
-        )
-
-        data class NameIdentifier(
-            val schemeUri: URI,
-            val nameIdentifier: String,
-            val nameIdentifierScheme: String
-        ) {
-            companion object {
-                fun fromORCID(orcid: String): NameIdentifier =
-                    NameIdentifier(
-                        schemeUri = URI.create("https://orcid.org"),
-                        nameIdentifier = "https://orcid.org/$orcid",
-                        nameIdentifierScheme = "ORCID"
-                    )
-            }
-        }
-
-        data class Title(
-            val title: String,
-            val lang: String = "en"
-        )
-
-        data class Subject(
-            val subject: String,
-            val lang: String = "en"
-        )
-
-        data class Type(
-            val resourceType: String,
-            val resourceTypeGeneral: String
-        )
-
-        data class RelatedIdentifier(
-            val relatedIdentifier: String,
-            val relatedIdentifierType: String,
-            val relationType: String = "References"
-        ) {
-            companion object {
-                fun fromDOI(doi: String): RelatedIdentifier =
-                    RelatedIdentifier(
-                        relatedIdentifier = doi,
-                        relatedIdentifierType = "DOI"
-                    )
-            }
-        }
-
-        data class Rights(
-            val rights: String,
-            val rightsUri: URI
-        ) {
-            companion object {
-                val CC_BY_SA_4_0 = Rights(
-                    rights = "Creative Commons Attribution-ShareAlike 4.0 International License.",
-                    rightsUri = URI.create("https://creativecommons.org/licenses/by-sa/4.0/")
-                )
-            }
-        }
-
-        data class Description(
-            val description: String,
-            val descriptionType: String = "Abstract"
-        )
-    }
-
-    private fun DoiService.RegisterCommand.toDataCiteRequest(prefix: String, action: String, clock: Clock): RegisterDoiRequest =
-        RegisterDoiRequest(
+    private fun DoiService.RegisterCommand.toDataCiteRequest(prefix: String, action: String, clock: Clock): DataCiteJson =
+        DataCiteJson(
             attributes = Attributes(
                 doi = "$prefix/$suffix",
                 event = action,