+++ graph/graph-application/src/main/kotlin/eu/tib/orkg/prototype/contenttypes/api/comparison-queries.kt
@@ -1,9 +1,9 @@
 package eu.tib.orkg.prototype.contenttypes.api
 
+import eu.tib.orkg.prototype.community.domain.model.ContributorId
 import eu.tib.orkg.prototype.contenttypes.domain.model.Comparison
 import eu.tib.orkg.prototype.contenttypes.domain.model.ComparisonRelatedFigure
 import eu.tib.orkg.prototype.contenttypes.domain.model.ComparisonRelatedResource
-import eu.tib.orkg.prototype.community.domain.model.ContributorId
 import eu.tib.orkg.prototype.statements.api.VisibilityFilter
 import eu.tib.orkg.prototype.statements.domain.model.ContributionInfo
 import eu.tib.orkg.prototype.statements.domain.model.ThingId
@@ -22,6 +22,8 @@ interface RetrieveComparisonUseCase {
     fun findAllRelatedResources(comparisonId: ThingId, pageable: Pageable): Page<ComparisonRelatedResource>
     fun findRelatedFigureById(comparisonId: ThingId, id: ThingId): Optional<ComparisonRelatedFigure>
     fun findAllRelatedFigures(comparisonId: ThingId, pageable: Pageable): Page<ComparisonRelatedFigure>
+    /* An unpublished comparison is a comparison that does not have a DOI and is not a draft comparison (ComparisonDraft) */
+    fun findAllCurrentListedAndUnpublishedComparisons(pageable: Pageable): Page<Comparison>
 }
 
 interface RetrieveComparisonContributionsUseCase {